Adds missing vm parameter to examples
Currently some of the examples don't work as they miss the vm
parameter meaning the task fails with: "Input validation failed.".
Fault detail is "No matching signature found, make sure that
mandatory attributes are provided.". HTTP response code is 400."

This commit adds the vm parameter which allows the task to
successfully complete and create a template.