roles: hosted_engine_setup: Improve engine logs fetching (#151)
* roles: hosted_engine_setup: Refactor add_host to own tasks file

* roles: hosted_engine_setup: Call 'sync' before fetching engine logs

This is an addition to [1], missing there, updating also
partial_execution.yml . That said:

- This does not (usually?) reproduce on CI, perhaps simply because we
run there with low memory (thus forcing the kernel to sync often)

- I do not know well which parts of this code run on the host, which on
the engine, and what and when adds the engine to the inventory, so not
sure it would always work.

Copying the commit message from [1]:

Sometimes, on failed deploy attempts, engine-logs* directories are
created but left empty. I am guessing that this might be because the
files there are still cached in memory, and we do not fetch from the
engine vm itself (using e.g. ssh) but from the host, directly accessing
the vm disk image. So if, before that, we run 'sync' on the engine vm,
we have better chances to successfully get the logs from the image.

[1] https://github.com/oVirt/ovirt-ansible-hosted-engine-setup/pull/325

Bug-Url: https://bugzilla.redhat.com/1844965

* roles: hosted_engine_setup: Collect logs also on engine-setup failure