roles: hosted_engine_setup: Add after_add_host hook (#181)
* roles: hosted_engine_setup: Add after_add_host hook
Allow running custom playbooks after trying to add the host to the
engine.

Bug-Url: https://bugzilla.redhat.com/1893385

* Add missing space
cut -d' ' also works, but both emacs and vi syntax-highlight this
wrongly, and IMO it's less common. See also e.g.:
https://stackoverflow.com/questions/10818443/short-long-options-with

* On 'Add host' failures, collect also warning events
If we failed due to a missing required network, this will be logged as a
warning, not error. Collect that as well.

* Pause if Add host failed

* Fix error message collection

* Fix typo

* Add required_networks_fix.yml example