roles: hosted_engine_setup: Collect all engine /var/log
In the code collecting stuff from the engine machine, we collect only
/var/log/messages and /var/log/ovirt-engine.

Change this to collect all of /var/log.

I think this is useful in general, and does not add too much waste (in
deploy time and space after that).

For now, I am pushing this to try to debug:
https://bugzilla.redhat.com/1915329