infra: fix storage target
The issue was that the role was trying to edit the storage connection
target when the storage domain has multiple connection targets.

This patch checks if the target which user input is in the list
of targets from the storage domain if so it will ignore it else it will edit
it.

https://bugzilla.redhat.com/show_bug.cgi?id=1943221

Signed-off-by: mnecas <necas.marty@gmail.com>