roles: hosted_engine_setup: Allow custom admin username
Make the username used in auth_sso.yml customizable.

Previous patches should set it based on whether keycloak is configured.

For this code to work with partial_execution.yml, it relies on  the
caller to pass the correct username in each call. This is done in a
currently-pending patch to ovirt-hosted-engine-setup.