Configure keycloak (#488)
* roles: hosted_engine_setup: Allow configuring keycloak

Please note that whether this actually causes engine-setup to configure
keycloak depends on availability and functionality of relevant code.
Right now this code is not included by default. This patch is in
preparation for changing this.

* roles: hosted_engine_setup: Set correct admin username

based on whether keycloak was configured or not.

Pass the result to otopi (ovirt-hosted-engine-setup).

* roles: hosted_engine_setup: Allow custom admin username

Make the username used in auth_sso.yml customizable.

Previous patches should set it based on whether keycloak is configured.

For this code to work with partial_execution.yml, it relies on  the
caller to pass the correct username in each call. This is done in a
currently-pending patch to ovirt-hosted-engine-setup.

* he_enable_keycloak is always defined, because it's specified in default vars

Co-authored-by: Martin Perina <mperina@redhat.com>