roles: hosted_engine_setup: Require 'detail: Up' (#498)
When checking, near the end of the deployment, if the engine is healthy,
do not only check that 'health' is 'good', but also that 'detail' is
'Up'.

We recently had a case in CI where due to some problem the engine took a
long time to start, timing out the monitoring in HA, causing it to stop
the VM, and 'detail' was 'Powering Down', but this went unnoticed,
causing things to fail much later. So, in such a case, fail early.