lint: he: Stop excluding
Run it on HE separately.

Exclude the test 'var-spacing':

- We have far too many places that break this test to fix all of them
automatically without dedicated language-specific tooling. I am not
aware of such tooling. I tried doing this with ad-hoc 'find | xargs sed'
hacks but this proved to be too hard to do safely.

- IMO this rule is more a matter of personal taste/style than actually
an improvement for the stability/robustness/correctness of the code.
While I personally do prefer the spacing it mandates, my preference does
not extend to spending time breaking working code and then fixing it.

Exclude 'ignore-errors':

- That's somewhat simpler to fix than above, but I do not see why. I
tried changing all 'ignore_errors: true' to 'failed_when: false', but
that's not enough - they are not exactly the same (and rightly so).

- If we think there is real value in 'ignore-errors', we can fix later,
no need to postpone upgrade of lint.

Exclude 'var-naming':

- It's breaking on variables like he_mem_size_MB and he_disk_size_GB,
which do make sense.

- Completely replacing all of them with lower letters will require
changes also in ovirt-hosted-engine-setup.

- If needed, can be done later.

Bug-Url: https://bugzilla.redhat.com/2097333