he: Add loop variables to include_tasks uses
Especially those that allow custom user-supplied hooks, risk confusing
the default variable "item" with loops inside such globs/hooks. So use
custom variable names.

This affects us in ovirt-system-tests, which has loops in
aa enginevm_before_engine_setup hook:

- Copy sysstat rpm package to HE VM
- Copy sysstat dependencies to HE VM

This caused these errors in the output:

    The loop variable 'item' is already in use

It was quite hard to understand the source of the errors, because they
appear only in the stderr of ansible-playbook, which in `hosted-engine
--deploy` is logged after the playbook is finished. I eventually found
these with a custom patch to the callback to log everything also to
stderr:

https://github.com/oVirt/ovirt-hosted-engine-setup/pull/54