HE: Filter IP addresses based on 'ipv6_deployment'
At the beginning of HE deployment we set the 'ipv6_deployment' fact
based on host's address. This var should be the source of truth for
whether to use an IPv4 or IPv6 address for the HE VM.

Until now we were simply picking up the first address that was reported
by 'virsh' and using this throughout the rest of the process.

This PR adds filtering of the reported addresses based on the value of
'ipv6_deployment'. Additionally we're picking the first address to make
other usages of 'local_vm_ip' nicer.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>