Fixing create VM from template with non admin user
This patch fixes creating a VM from a template where the template
resides in the Default DC.
REST API has a problem of not getting the href for the DC when dealing
with Default cluster.
In order to fix that a special treatment was added to get the right
Default DC object.
Solution will not work if the Default DC/Cluster name was changed from
"Default" to something else. ( we don't have any indication apart of the
name for Default DC/Cluster)

Signed-Off-By: <Eli Mesika: emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2078946