Fix error around using dotfile in changelog fragments
Previously ansible lint ignored dotfiles inside `changelog/fragments`
directory. Ast we need to have at least one file in this directory
otherwise it will not be created when cloning from git.
Now ansible lint raises an error when `.placeholder` file exists in
`changelog/fragments`, so we need to rename it to `.keep` to silence
this error.

Signed-off-by: Martin Perina <mperina@redhat.com>