Debug appliance virt install failure (#664)
* roles: hosted_engine_setup: Log the Local VM console

Signed-off-by: Yedidyah Bar David <didi@redhat.com>

* WIP: Help debug bootstrap VM boot issues

- --graphics none: Will prevent creating a vnc console. I think that's
ok, at this stage. This should not affect the eventual engine VM conf.
So, use the serial console as the main one.

- --autoconsole text: Causes virt-install to automatically connect to
the VM's console and show it. This should make the console output be
included in the ansible log.

- --boot hd,bootmenu.enable=on,bios.useserial=on: Allow getting the boot
menu, useful for debugging/investigation. Also adds a few seconds to the
boot process.

Signed-off-by: Yedidyah Bar David <didi@redhat.com>

Signed-off-by: Yedidyah Bar David <didi@redhat.com>