Support boot disk renaming and resizing.
Renames the boot disk when the `boot_disk_name` variable is defined
within the profile or the VM properties. (The latter has precedence
over the former). The VM disk must have the bootable flag set.

By default, if `boot_disk_name` is defined, the VM name will be
prepended to the boot disk name so the final name would be:

`{{ vm_name }}_{{ boot_disk_name }}`

This behavior can be disabled by setting the variable
`boot_disk_use_vm_name_prefix` to `false`. In this case the
final name of the boot disk will be the value of `boot_disk_name`.

The boot disk can also be resized by defining the `boot_disk_size`
variable.

The value can contain a suffix that complies to the IEC 60027-2
standard (for example 10GiB, 1024MiB)

Signed-off-by: Miguel Martín <mmartinv@redhat.com>