image_template: enable image template version name (#733)
* image_template: allow to setup template version name

Allow to create templates with the same name but with
different version by setting up the `template_version_name`
variable. Creating a new version requires the base template
to exist. If not specified the value of the variable is
`base version` which is the default value for base templates.

Signed-off-by: Miguel Martín <mmartinv@redhat.com>

* image_template: allow to setup template OS and description

Allow to setup the template Operative System and the
Description.

Signed-off-by: Miguel Martín <mmartinv@redhat.com>

* Add changelog for #733

Signed-off-by: Miguel Martín <mmartinv@redhat.com>

---------

Signed-off-by: Miguel Martín <mmartinv@redhat.com>