Not possible to attach LUN to running VM through ansible
Issue: Ansible example shows that lun can be attached to vm. The example shows to add lun with lun_id and storage_type attribute. The ansibles scripts fails to execute as it fails to attach disk because it failed to find the lun and associated disk info. The issue was the lun directory structure has lun_id attribute and the lookup code was using just id instead of lun_id.

https://docs.ansible.com/ansible/latest/collections/ovirt/ovirt/ovirt_disk_module.html#parameter-logical_unit/lun_id

Fix: Fix the attribute name that used to get the disk info.

Signed-off-by: Saksham Srivastava <saksham.sa.srivastava@oracle.com>