Fix missing cluster name when starting VM
Issue:
When the VM gets started on specific cluster with the vm_infra role it
fails with error:
```
Fault reason is "Incomplete parameters". Fault detail is "Vm [cluster.id|name] required for add". HTTP response code is 400.
```

This is caused because the role does not pass the cluster information to
the ovirt_vm module.

Thank you @gsigris for finidng this issue!

Signed-off-by: Martin Necas <mnecas@redhat.com>